Fixed the test failure issue.
- Separated the tf and pytorch dependencies, we don't need to install
one if not required.
- Only the stable pipeline will be enabled as default, other one will be
used at dev stage only.
- Migrated the requirements and docker files under a folder.
