+++ source/pic2card/docker/Dockerfile-tf_serving
