+++ source/pic2card/requirements/requirements-frozen_graph.txt
