+++ source/pic2card/requirements/requirements-pytorch.txt
@@ -1,0 +1,3 @@
+torch==1.5.0
+torchvision==0.6.0
+detecto==1.1.4
