+++ source/pic2card/requirements/requirements.txt
@@ -1,5 +1,4 @@
-# tf serving specific dependencies.
-# avoids the tensorflow core library.
+# DL library dependencies are kept eparate due to its size.
 flask==1.1.1
 flask_restplus==0.13.0
 flask-cors==3.0.8
